+++ src/sources.rs
@@ -8,8 +8,8 @@ use reqwest::StatusCode;
 use crate::cache::{Error, Result};
 
 pub trait ConfigSource<S> {
-    fn fetch(&self) -> Result<(u64, S)>;
-    fn fetch_if_newer(&self, version: u64) -> Result<Option<(u64, S)>>;
+    fn fetch(&self) -> Result<(u128, S)>;
+    fn fetch_if_newer(&self, version: u128) -> Result<Option<(u128, S)>>;
 }
 
 pub struct HttpConfigSource {
@@ -27,7 +27,7 @@ impl HttpConfigSource {
 }
 
 impl ConfigSource<Response> for HttpConfigSource {
-    fn fetch(&self) -> Result<(u64, Response)> {
+    fn fetch(&self) -> Result<(u128, Response)> {
         let fetched = self.fetch_if_newer(0)?;
         match fetched {
             None => Err(Error::new("Unconditional fetch returned nothing")),
@@ -35,8 +35,8 @@ impl ConfigSource<Response> for HttpConfigSource {
         }
     }
 
-    fn fetch_if_newer(&self, version: u64) -> Result<Option<(u64, Response)>> {
-        let date = UNIX_EPOCH.add(Duration::from_millis(version));
+    fn fetch_if_newer(&self, version: u128) -> Result<Option<(u128, Response)>> {
+        let date = UNIX_EPOCH.add(Duration::from_millis(version as u64));
         let resp = self.client.get(self.url.as_str())
             .header("If-Modified-Since", httpdate::fmt_http_date(date))
             .send()?;
@@ -44,7 +44,7 @@ impl ConfigSource<Response> for HttpConfigSource {
         if resp.status().is_success() {
             let version = if let Some(header) = resp.headers().get("Last-Modified") {
                 let date = httpdate::parse_http_date(header.to_str()?)?;
-                date.duration_since(UNIX_EPOCH)?.as_millis() as u64
+                date.duration_since(UNIX_EPOCH)?.as_millis()
             } else {
                 0
             };
@@ -71,22 +71,21 @@ impl<P: AsRef<Path>> LocalFileConfigSource<P> {
 }
 
 impl<P: AsRef<Path>> ConfigSource<BufReader<File>> for LocalFileConfigSource<P> {
-    fn fetch(&self) -> Result<(u64, BufReader<File>)> {
+    fn fetch(&self) -> Result<(u128, BufReader<File>)> {
         match self.fetch_if_newer(0)? {
             None => Err(Error::new("Unconditional fetch failed")),
             Some((v, b)) => Ok((v, b))
         }
     }
 
-    fn fetch_if_newer(&self, version: u64) -> Result<Option<(u64, BufReader<File>)>> {
+    fn fetch_if_newer(&self, version: u128) -> Result<Option<(u128, BufReader<File>)>> {
         let file = File::open(&self.path)?;
         let metadata = file.metadata()?;
         match metadata.modified() {
             Ok(t) => {
                 let mtime = t.duration_since(UNIX_EPOCH)?.as_millis();
-                //If version overflows a u64 fetch_if_newer will be effectively unconditional
-                if (version as u128) < mtime {
-                    Ok(Some((mtime as u64, BufReader::new(file))))
+                if version < mtime {
+                    Ok(Some((mtime, BufReader::new(file))))
                 } else {
                     Ok(None)
                 }
@@ -95,7 +94,6 @@ impl<P: AsRef<Path>> ConfigSource<BufReader<File>> for LocalFileConfigSource<P> 
             //We're on a platform that doesn't support file mtime, unconditional it is.
             Err(_) => Ok(Some((0, BufReader::new(file))))
         }
-
     }
 }
